Refactor renderer
