+++ compositor_render/src/transformations/builtin.rs
@@ -1,9 +1,16 @@
-use compositor_common::scene::{
-    builtin_transformations::{BuiltinSpec, TransformToResolutionStrategy},
-    Resolution,
+use std::sync::Arc;
+
+use compositor_common::{
+    renderer_spec::FallbackStrategy,
+    scene::{
+        builtin_transformations::{BuiltinSpec, TransformToResolutionStrategy},
+        Resolution,
+    },
 };
 
 use crate::utils::rgba_to_wgpu_color;
+
+use super::shader_executor::ShaderExecutor;
 
 mod box_layout;
 pub mod error;
@@ -12,12 +19,15 @@ pub mod params;
 pub mod transformations;
 pub mod utils;
 
-#[derive(Debug, Clone)]
-pub struct Builtin(pub BuiltinSpec);
+#[derive(Debug)]
+pub struct Builtin {
+    pub spec: BuiltinSpec,
+    pub executor: Arc<ShaderExecutor>,
+}
 
 impl Builtin {
     pub fn clear_color(&self) -> Option<wgpu::Color> {
-        match &self.0 {
+        match &self.spec {
             BuiltinSpec::TransformToResolution { strategy, .. } => match strategy {
                 TransformToResolutionStrategy::Stretch | TransformToResolutionStrategy::Fill => {
                     None
@@ -52,7 +62,7 @@ impl Builtin {
                 })
         }
 
-        match self.0 {
+        match self.spec {
             BuiltinSpec::TransformToResolution { resolution, .. } => resolution,
             BuiltinSpec::FixedPositionLayout { resolution, .. } => resolution,
             BuiltinSpec::TiledLayout { resolution, .. } => resolution,
@@ -62,12 +72,22 @@ impl Builtin {
     }
 
     pub fn resolution_from_spec(&self) -> Option<Resolution> {
-        match self.0 {
+        match self.spec {
             BuiltinSpec::TransformToResolution { resolution, .. } => Some(resolution),
             BuiltinSpec::FixedPositionLayout { resolution, .. } => Some(resolution),
             BuiltinSpec::TiledLayout { resolution, .. } => Some(resolution),
             BuiltinSpec::MirrorImage { .. } => None,
             BuiltinSpec::CornersRounding { .. } => None,
+        }
+    }
+
+    pub fn fallback_strategy(&self) -> FallbackStrategy {
+        match self.spec {
+            BuiltinSpec::TransformToResolution { .. }
+            | BuiltinSpec::FixedPositionLayout { .. }
+            | BuiltinSpec::TiledLayout { .. }
+            | BuiltinSpec::MirrorImage { .. }
+            | BuiltinSpec::CornersRounding { .. } => FallbackStrategy::FallbackIfAllInputsMissing,
         }
     }
 }
