+++ compositor_render/src/transformations/builtin/error.rs
@@ -1,4 +1,4 @@
-use crate::transformations::shader::CreateShaderError;
+use crate::transformations::shader_executor::CreateShaderError;
 
 #[derive(Debug, thiserror::Error)]
 pub enum InitBuiltinError {
