+++ compositor_render/src/transformations/web_renderer/node.rs
@@ -58,7 +58,7 @@ impl WebRendererNode {
     }
 
     pub fn fallback_strategy(&self) -> FallbackStrategy {
-        self.renderer.fallback_strategy
+        self.renderer.fallback_strategy()
     }
 
     fn ensure_buffer_size(
